refactoring app.js completed
