+++ client/src/components/ToastContainerCom/ToastContainerCom.js
@@ -1,0 +1,21 @@
+import React from "react";
+import { ToastContainer } from "react-toastify";
+
+function ToastContainerCom() {
+  return (
+    <ToastContainer
+      position="top-center"
+      autoClose={3000}
+      hideProgressBar
+      newestOnTop={false}
+      closeOnClick
+      rtl={false}
+      pauseOnFocusLoss
+      draggable
+      pauseOnHover
+      theme="light"
+    />
+  );
+}
+
+export default ToastContainerCom;
